requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
mssql-python>=1.0.0
python-dotenv>=1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator>=2.0.0
